move core function to a more central place
